add notebook to test wosis
This notebook contains a test of wosis and metaknowledge, the idea was
to use wosis to make topic analysis and try some network analysis.
They were working in a first test, but latter I had some trouble when
reading the initial txt file and I was not able to fix it. I am saving
this notebook as it is to try to solve the problem later.