add notebook with topic analysis
I want to try using topic analysis as a way to identify clusters in the
meta-review component of the literature review. Here I add a test using
litstudy to make this topic analysis. Currently it is using the scopus
API and the data need to be dowloaded each time I want to use it.