Add initial files for modelling analysis
Including first version of deduplication and topic modelling. This files
are similar to the meta-review, but the have new features, such as: the
deduplication makes unification of journal names and plot most present
journals, topic modelling have different parameters, plot quantification
of papers in journals and export papers per journal.