Change visual of result plots
- Change color paletter of percentage of metric category per topic
model, predefining a palette before hand so that in all plots each
category has the same color
- Change diversity-based plot to annotate both the number of metrics and
the total number of metrics in the block. Also color is refering to the
percentual value. Both percentual and absolute number tell different and
relavant things, so now the plot shows both information.
- Do not remove columns when reading the dataframe, mainly for
simplicity and to facilitate analysing additional aspects if needed.