Change review diversity plot to include denominator always
In the review diversity figure we were including and annotation of the
denominator only when the numberator is different from zero. However,
even when the value is zero, it still can shows a relevant information.
So it the denominator value always, even when the numerator is zero.

Also, it adds the y-axis label for the bar plots figures.