Adds validation and updates parameters for modelling review
Similar to what was done in the meta-review, to improve the previous
analysis and answer to the reviewers comments, this commits adds a more
systematic validation on the number of topics in the modelling review
component. We use topic coherence and diversity as quality metrics.
Here we also updated the selected parameters and removed the journal
selction method.